feat(ui): add `RadioGroup` and `RadioItem` to the Dialog component
