+++ packages/ui/src/Dialog/Context.tsx
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+/** Internal use only. */
+export const DialogContext = createContext<{ showCloseButton: boolean }>({
+  showCloseButton: true,
+});
