+++ packages/ui/src/Dialog/index.tsx
@@ -1,110 +1,11 @@
-import { createContext, ReactNode, useContext } from 'react';
-import { styled } from 'styled-components';
+import { ReactNode } from 'react';
 import * as RadixDialog from '@radix-ui/react-dialog';
-import { Text } from '../Text';
-import { X } from 'lucide-react';
-import { ButtonGroup, ButtonGroupProps } from '../ButtonGroup';
-import { Button } from '../Button';
-import { Density } from '../Density';
-import { Display } from '../Display';
-import { Grid } from '../Grid';
-import { MotionProp } from '../utils/MotionProp';
-import { motion } from 'framer-motion';
-
-const Overlay = styled(RadixDialog.Overlay)`
-  backdrop-filter: blur(${props => props.theme.blur.xs});
-  background-color: ${props => props.theme.color.other.overlay};
-  position: fixed;
-  inset: 0;
-  z-index: auto;
-`;
-
-const FullHeightWrapper = styled.div`
-  height: 100%;
-  min-height: 100svh;
-  max-height: 100lvh;
-  position: relative;
-
-  display: flex;
-  align-items: center;
-`;
-
-/**
- * We make a full-screen wrapper around the dialog's content so that we can
- * correctly position it using the same `<Display />`/`<Grid />` as the
- * underlying page uses. Note that we use a `styled.div` here, rather than
- * `styled(RadixDialog.Content)`, because Radix adds an inline `pointer-events:
- * auto` style to that element. We need to make sure there _aren't_ pointer
- * events on the dialog content, because of the aforementioned full-screen
- * wrapper that appears over the `<Overlay />`. We want to make sure that clicks
- * on the full-screen wrapper pass through to the underlying `<Overlay />`, so
- * that the dialog closes when the user clicks there.
- */
-const DialogContent = styled.div<{ $zIndex?: number }>`
-  position: fixed;
-  inset: 0;
-  pointer-events: none;
-  ${props => props.$zIndex && `z-index: ${props.$zIndex};`}
-`;
-
-const DialogContentCard = styled(motion.div)`
-  position: relative;
-  width: 100%;
-  max-height: 75%;
-  box-sizing: border-box;
-
-  background: ${props => props.theme.color.other.dialogBackground};
-  border: 1px solid ${props => props.theme.color.other.tonalStroke};
-  border-radius: ${props => props.theme.borderRadius.xl};
-  backdrop-filter: blur(${props => props.theme.blur.xl});
-
-  display: flex;
-  flex-direction: column;
-
-  /**
-   * We add 'pointer-events: auto' here so that clicks _inside_ the content card
-   * work, even though the _outside_ clicks pass through to the underlying
-   * '<Overlay />'.
-   */
-  pointer-events: auto;
-`;
-
-const DialogChildrenWrap = styled.div`
-  overflow-y: auto;
-
-  display: flex;
-  flex-direction: column;
-  gap: ${props => props.theme.spacing(6)};
-
-  padding-bottom: ${props => props.theme.spacing(8)};
-  padding-left: ${props => props.theme.spacing(6)};
-  padding-right: ${props => props.theme.spacing(6)};
-`;
-
-const DialogHeader = styled.header`
-  position: sticky;
-  top: 0;
-
-  display: flex;
-  flex-direction: column;
-  gap: ${props => props.theme.spacing(4)};
-  color: ${props => props.theme.color.text.primary};
-
-  padding-top: ${props => props.theme.spacing(8)};
-  padding-bottom: ${props => props.theme.spacing(6)};
-  padding-left: ${props => props.theme.spacing(6)};
-  padding-right: ${props => props.theme.spacing(6)};
-`;
-
-/**
- * Opening the dialog focuses the first focusable element in the dialog. That's why the Close button
- * should be positioned absolutely and rendered as the last element in the dialog content.
- */
-const DialogClose = styled.div`
-  position: absolute;
-  top: ${props => props.theme.spacing(8)};
-  right: ${props => props.theme.spacing(6)};
-`;
+import { DialogContext } from './Context.tsx';
+import { EmptyContent, DialogEmptyContentProps } from './EmptyContent.tsx';
+import { Content, DialogContentProps } from './Content.tsx';
+import { Trigger, DialogTriggerProps } from './Trigger.tsx';
+import { RadioGroup, DialogRadioGroupProps } from './RadioItemGroup.tsx';
+import { RadioItem, DialogRadioItemProps } from './RadioItem';
 
 interface ControlledDialogProps {
   /**
@@ -235,119 +136,16 @@ export const Dialog = ({ children, onClose, isOpen }: DialogProps) => {
   );
 };
 
-export interface DialogEmptyContentProps {
-  children?: ReactNode;
-  /** @deprecated this prop will be removed in the future */
-  zIndex?: number;
-}
-
-const EmptyContent = ({ children, zIndex }: DialogEmptyContentProps) => {
-  return (
-    <RadixDialog.Portal>
-      <Overlay />
-
-      <RadixDialog.Content>
-        <DialogContent $zIndex={zIndex}>{children}</DialogContent>
-      </RadixDialog.Content>
-    </RadixDialog.Portal>
-  );
-};
 Dialog.EmptyContent = EmptyContent;
-
-/** Internal use only. */
-const DialogContext = createContext<{ showCloseButton: boolean }>({
-  showCloseButton: true,
-});
-
-export interface DialogContentProps<IconOnlyButtonGroupProps extends boolean | undefined>
-  extends MotionProp {
-  children?: ReactNode;
-  /** Renders the element after the dialog title. These elements will be sticky to the top of the dialog */
-  headerChildren?: ReactNode;
-  title: string;
-  /**
-   * If you want to render CTA buttons in the dialog footer, use
-   * `buttonGroupProps`. The dialog will then render a `<ButtonGroup />` using
-   * these props.
-   */
-  buttonGroupProps?: IconOnlyButtonGroupProps extends boolean
-    ? ButtonGroupProps<IconOnlyButtonGroupProps>
-    : undefined;
-  /** @deprecated this prop will be removed in the future */
-  zIndex?: number;
-}
-
-const Content = <IconOnlyButtonGroupProps extends boolean | undefined>({
-  children,
-  headerChildren,
-  title,
-  buttonGroupProps,
-  motion,
-  zIndex,
-}: DialogContentProps<IconOnlyButtonGroupProps>) => {
-  const { showCloseButton } = useContext(DialogContext);
-
-  return (
-    <EmptyContent zIndex={zIndex}>
-      <Display>
-        <Grid container>
-          <Grid mobile={0} tablet={2} desktop={3} xl={4} />
-
-          <Grid mobile={12} tablet={8} desktop={6} xl={4}>
-            <FullHeightWrapper>
-              <DialogContentCard {...motion}>
-                <DialogHeader>
-                  <RadixDialog.Title asChild>
-                    <Text xxl as='h2'>
-                      {title}
-                    </Text>
-                  </RadixDialog.Title>
-                  {headerChildren}
-                </DialogHeader>
-
-                <DialogChildrenWrap>
-                  {children}
-
-                  {buttonGroupProps && <ButtonGroup {...buttonGroupProps} column />}
-                </DialogChildrenWrap>
-
-                {showCloseButton && (
-                  <Density compact>
-                    <RadixDialog.Close asChild>
-                      <DialogClose>
-                        <Button icon={X} iconOnly priority='secondary'>
-                          Close
-                        </Button>
-                      </DialogClose>
-                    </RadixDialog.Close>
-                  </Density>
-                )}
-              </DialogContentCard>
-            </FullHeightWrapper>
-          </Grid>
-
-          <Grid mobile={0} tablet={2} desktop={3} xl={4} />
-        </Grid>
-      </Display>
-    </EmptyContent>
-  );
-};
 Dialog.Content = Content;
-
-export interface DialogTriggerProps {
-  children: ReactNode;
-  /**
-   * Change the default rendered element for the one passed as a child, merging
-   * their props and behavior.
-   *
-   * Uses Radix UI's `asChild` prop under the hood.
-   *
-   * @see https://www.radix-ui.com/primitives/docs/guides/composition
-   */
-  asChild?: boolean;
-}
-
-const Trigger = ({ children, asChild }: DialogTriggerProps) => (
-  <RadixDialog.Trigger asChild={asChild}>{children}</RadixDialog.Trigger>
-);
 Dialog.Trigger = Trigger;
+Dialog.RadioGroup = RadioGroup;
+Dialog.RadioItem = RadioItem;
+
+export type {
+  DialogTriggerProps,
+  DialogEmptyContentProps,
+  DialogContentProps,
+  DialogRadioGroupProps,
+  DialogRadioItemProps,
+};
